// src/api/generated/models/communicationOptOutWithRelations.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */
import type { CommunicationOptOutWithRelationsChannel } from './communicationOptOutWithRelationsChannel';
import type { Person } from './person';

/**
 * (tsType: CommunicationOptOutWithRelations, schemaOptions: { includeRelations: true })
 */
export interface CommunicationOptOutWithRelations {
  id?: string;
  channel: CommunicationOptOutWithRelationsChannel;
  /** @nullable */
  reason?: string | null;
  createdAt: string;
  updatedAt: string;
  personId: string;
  person?: Person;
  foreignKey?: unknown;
}